update in analysis window
This is an updated version of the preregistration of 20 June 2022. The analysis window of the Affective state adjective segment is now 0-1.5s after the onset of the segment. The previous preregistration was 0-1s, but in the preregistration of Li Kloostra's thesis project we already decided to extend the analysis to 1.5s after the onset of the segment. In addition, we will also use Python to do part of the analyses. This amendment is defined prior to the data analysis.